Fix SecurityProtocol import after upstream move (#642)
Merging to fix downstream builds. Tested locally with the latest Kafka trunk.
