+++ core/src/test/java/io/confluent/kafka/schemaregistry/ClusterTestHarness.java
@@ -19,7 +19,7 @@ import io.confluent.common.utils.IntegrationTest;
 import kafka.utils.CoreUtils;
 import org.I0Itec.zkclient.ZkClient;
 import org.apache.kafka.common.network.ListenerName;
-import org.apache.kafka.common.protocol.SecurityProtocol;
+import org.apache.kafka.common.security.auth.SecurityProtocol;
 import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.common.utils.Utils;
 import org.junit.After;
