+++ core/src/test/java/io/confluent/kafka/schemaregistry/SASLClusterTestHarness.java
@@ -21,7 +21,7 @@ import kafka.server.KafkaConfig;
 import kafka.utils.JaasTestUtils;
 import kafka.utils.TestUtils;
 import org.apache.kafka.common.config.SaslConfigs;
-import org.apache.kafka.common.protocol.SecurityProtocol;
+import org.apache.kafka.common.security.auth.SecurityProtocol;
 import org.apache.kafka.common.security.authenticator.LoginManager;
 import org.junit.After;
 import org.junit.Before;
@@ -31,7 +31,6 @@ import scala.Option;
 import scala.collection.JavaConversions;
 import scala.collection.Seq;
 import scala.collection.immutable.List;
-import scala.collection.mutable.Buffer;
 
 import javax.security.auth.login.Configuration;
 import java.io.File;
