+++ core/src/test/java/io/confluent/kafka/schemaregistry/storage/KafkaStoreTest.java
@@ -18,11 +18,7 @@ package io.confluent.kafka.schemaregistry.storage;
 import io.confluent.kafka.schemaregistry.rest.SchemaRegistryConfig;
 import kafka.admin.AdminUtils;
 import kafka.admin.RackAwareMode;
-import kafka.cluster.Broker;
 import kafka.log.LogConfig;
-import org.apache.kafka.common.config.ConfigException;
-import org.apache.kafka.common.network.ListenerName;
-import org.apache.kafka.common.protocol.SecurityProtocol;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -33,10 +29,6 @@ import io.confluent.kafka.schemaregistry.ClusterTestHarness;
 import io.confluent.kafka.schemaregistry.storage.exceptions.StoreException;
 import io.confluent.kafka.schemaregistry.storage.exceptions.StoreInitializationException;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 import java.util.Properties;
 
 import static org.junit.Assert.assertEquals;
